Split deployment logs tests by the file that is being tested
